test: execute Twemoji description behavior
