+++ AGENTS.md
@@ -32,7 +32,7 @@
 
 ## Testing guidance
 
-- No XCTest target is present; `make check` is the maintained static regression gate and also parses the Xcode project when `xcodebuild` is available.
+- No XCTest target is present; `make check` combines the maintained static regression gate with executable Twemoji description behavior tests when `swiftc` is available, and also parses the Xcode project when `xcodebuild` is available.
 - `make check` and `make xcode-build` do not prove Messages-host interaction; record the exact Xcode/iOS runtime and separately exercise sticker selection, preview, send, transcript, and VoiceOver behavior.
 - Start with the narrowest relevant test or Make target, then run `make check` before handing off if the change is not documentation-only.
 - Keep README verification notes in sync when commands, fixtures, or supported toolchains change.
