+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 2026-06-16
+
+- Extracted Twemoji description behavior into framework-independent Swift and
+  added executable valid-decoding and invalid-input fallback cases to
+  `make check`.
+
 ## 2026-06-15
 
 - Bundled Twemoji graphics are attributed in THIRD_PARTY_NOTICES.md under CC BY 4.0.
